refactor: Improve Prometheus Health Check in Tests (#25)
## The Issue

This pull request addresses an issue where the Prometheus health check
in our tests was not sufficiently validating the service's
functionality. The previous check only verified the presence of a
specific string in the response from the `/query` endpoint, which wasn't
a reliable indicator of Prometheus's health or its ability to collect
and expose metrics.

<!-- Provide a brief description of the issue. -->

## How This PR Solves The Issue

This change replaces the old check with two more robust checks:

1. **API Availability:** Verifies that the Prometheus API is available
by querying the `/api/v1/status/config` endpoint and checking for a
successful response with `\"status\":\"success\"`.
2. **Metrics Exposure:** Confirms that Prometheus is exposing metrics by
querying the `/metrics` endpoint and checking for the presence of the
`TYPE prometheus_build_info` metric.

## Automated Testing Overview

<!-- Please describe the tests introduced by this PR, or explain why no
tests are needed. -->

## Release/Deployment Notes

<!-- Does this affect anything else or have ramifications for other
code? Does anything have to be done on deployment? -->

---------

Co-authored-by: tyler36 <7234392+tyler36@users.noreply.github.com>