refactor: MySQL exporter (#27)
## The Issue

The `mysqld-exporter` name is misleading, as it suggests it's only
compatible with `mysqld` when it can be used with other MySQL-compatible
databases.

<!-- Provide a brief description of the issue. -->

## How This PR Solves The Issue

This pull request renames the `mysqld-exporter` service and related
files to `mysql-exporter` for consistency and clarity. The previous name
implied a specific MySQL daemon (mysqld), while the exporter can work
with various MySQL-compatible databases.

## Manual Testing Instructions

```bash
ddev add-on get https://github.com/tyler36/ddev-site-metrics/tarball/<branch>
ddev restart
```

## Automated Testing Overview

<!-- Please describe the tests introduced by this PR, or explain why no
tests are needed. -->

## Release/Deployment Notes

<!-- Does this affect anything else or have ramifications for other
code? Does anything have to be done on deployment? -->

---------

Co-authored-by: tyler36 <7234392+tyler36@users.noreply.github.com>