refactor: use LOKI_HTTP_PORT for Loki container (#41)
## The Issue

Loki port is hardcoded in serveral places. This makes it difficult to
update.

## How This PR Solves The Issue

This PR uses an ENV value as port number. It defaults to `3100`.

## Manual Testing Instructions

```bash
ddev add-on get https://github.com/tyler36/ddev-site-metrics/tarball/loki-use-env-port
ddev restart
```

## Automated Testing Overview

<!-- Please describe the tests introduced by this PR, or explain why no
tests are needed. -->

## Release/Deployment Notes

<!-- Does this affect anything else or have ramifications for other
code? Does anything have to be done on deployment? -->

---------

Co-authored-by: tyler36 <7234392+tyler36@users.noreply.github.com>