fix: node exporter targets host system (#51)
## The Issue

The node_exporter is designed to monitor the host system. Deploying in
containers requires extra care in order to avoid monitoring the
container itself.

## How This PR Solves The Issue

This PR attempts to follow the advise as given in
https://github.com/prometheus/node_exporter#docker

In my prefered workflow, this PR results in node_exporter displaying
stats for WSL (such as uptime).

## Manual Testing Instructions

```bash
ddev add-on get https://github.com/tyler36/ddev-site-metrics/tarball/<branch>
ddev restart
```

## Automated Testing Overview

<!-- Please describe the tests introduced by this PR, or explain why no
tests are needed. -->

## Release/Deployment Notes

<!-- Does this affect anything else or have ramifications for other
code? Does anything have to be done on deployment? -->

Co-authored-by: tyler36 <7234392+tyler36@users.noreply.github.com>