add 'ddev-generated' signature (#53)
## The Issue

- #<issue number>

<!-- Provide a brief description of the issue. -->

## How This PR Solves The Issue

This PR updates the Nginx dashboard to the one provided by the include
nginx-exporter.

It adds a `db` tag, to match the included MySQL dashboard.
In additional, a "#ddev-generated" signature is included to allow the
addon to manage the file.

Available at:
https://github.com/nginx/nginx-prometheus-exporter/blob/main/grafana/dashboard.json


## Manual Testing Instructions

```bash
ddev add-on get https://github.com/tyler36/ddev-site-metrics/tarball/<branch>
ddev restart
```

## Automated Testing Overview

<!-- Please describe the tests introduced by this PR, or explain why no
tests are needed. -->

## Release/Deployment Notes

<!-- Does this affect anything else or have ramifications for other
code? Does anything have to be done on deployment? -->

---------

Co-authored-by: tyler36 <7234392+tyler36@users.noreply.github.com>