feat: Add Prometheus config reload command (#54)
## The Issue

Currently, modifying the Prometheus configuration requires restarting
the entire ddev environment or manually restarting the Prometheus
container. This can be disruptive and time-consuming, especially during
development and testing.

## How This PR Solves The Issue

This pull request adds a `-r` or `--reload` flag to the `ddev
prometheus` command. When this flag is used, the command sends a POST
request to the Prometheus `/-/reload` endpoint, which triggers a
configuration reload. This allows changes to the Prometheus
configuration to be applied immediately without restarting the
container.

## Manual Testing Instructions

```bash
ddev add-on get https://github.com/tyler36/ddev-site-metrics/tarball/<branch>
ddev restart
```

## Automated Testing Overview

<!-- Please describe the tests introduced by this PR, or explain why no
tests are needed. -->

## Release/Deployment Notes

<!-- Does this affect anything else or have ramifications for other
code? Does anything have to be done on deployment? -->

---------

Co-authored-by: tyler36 <7234392+tyler36@users.noreply.github.com>