refactor: be explicit with scrape metric paths (#60)
## The Issue

This pull request addresses an issue where the `metrics_path` was not
explicitly defined in the scrape configurations.

While this did not seem to affect functionality, explicitly stating the
endpoint is considered good practice and helps new developers unfamiliar
with the project.

## How This PR Solves The Issue

This pull request addresses an issue where the `metrics_path` was not
explicitly defined in the scrape configurations for `node-exporter`,
`postgres-exporter`, and `prometheus` jobs in the Prometheus
configuration.

**Changes:**

- **prometheus/scrapers/node-exporter.yml:** Added `metrics_path:
'/metrics'` to the `node-exporter` job configuration.
- **prometheus/scrapers/postgres-exporter.yml:** Added `metrics_path:
'/metrics'` to the `postgres` job configuration.
- **prometheus/scrapers/prometheus.yml:** Added `metrics_path:
'/metrics'` to the `prometheus` job configuration.

## Manual Testing Instructions

```bash
ddev add-on get https://github.com/tyler36/ddev-site-metrics/tarball/<branch>
ddev restart
```

## Automated Testing Overview

<!-- Please describe the tests introduced by this PR, or explain why no
tests are needed. -->

## Release/Deployment Notes

<!-- Does this affect anything else or have ramifications for other
code? Does anything have to be done on deployment? -->

Co-authored-by: tyler36 <7234392+tyler36@users.noreply.github.com>