docs: update README to display used ports (#61)
## The Issue

This projects uses many ports internally, and exposes a few to the host.
While this addon uses the default ports for each service, it can be
difficult to keep track of what they are.

## How This PR Solves The Issue

This PR adds a table to the docs lsting ports used.

## Manual Testing Instructions

```bash
ddev add-on get https://github.com/tyler36/ddev-site-metrics/tarball/<branch>
ddev restart
```

## Automated Testing Overview

<!-- Please describe the tests introduced by this PR, or explain why no
tests are needed. -->

## Release/Deployment Notes

<!-- Does this affect anything else or have ramifications for other
code? Does anything have to be done on deployment? -->

---------

Co-authored-by: tyler36 <7234392+tyler36@users.noreply.github.com>